move definitions of functions into maphistory.cpp
wasn't planned i was looking in the wrong maphistory lol
